Rename Project Package (#39)
* #3878 Rename Project Package https://github.com/adempiere/adempiere/pull/3878
- Change the utils name for Service

* #3878 Rename Project Package https://github.com/adempiere/adempiere/pull/3878
- Rename the method the java standard
